Move SpanService impls into otel module
