+++ embrace-test-fakes/src/main/kotlin/io/embrace/android/embracesdk/fakes/FakeEmbraceSpanFactory.kt
@@ -1,0 +1,26 @@
+package io.embrace.android.embracesdk.fakes
+
+import io.embrace.android.embracesdk.internal.otel.schema.EmbType
+import io.embrace.android.embracesdk.internal.otel.spans.EmbraceSdkSpan
+import io.embrace.android.embracesdk.internal.otel.spans.EmbraceSpanFactory
+import io.embrace.android.embracesdk.internal.otel.spans.OtelSpanBuilderWrapper
+import io.embrace.android.embracesdk.spans.AutoTerminationMode
+import io.embrace.android.embracesdk.spans.EmbraceSpan
+
+class FakeEmbraceSpanFactory(
+    private val spanToReturn: EmbraceSdkSpan = FakeEmbraceSdkSpan.stopped()
+) : EmbraceSpanFactory {
+    override fun create(
+        name: String,
+        type: EmbType,
+        internal: Boolean,
+        private: Boolean,
+        parent: EmbraceSpan?,
+        autoTerminationMode: AutoTerminationMode,
+    ): EmbraceSdkSpan = spanToReturn
+
+    override fun create(
+        otelSpanBuilderWrapper: OtelSpanBuilderWrapper,
+        autoTerminationMode: AutoTerminationMode
+    ): EmbraceSdkSpan = spanToReturn
+}
